Components EffectUnit refactor to solve some bugs
Move some initialization code into EffectUnit.setCurrentUnit and call
setCurrentUnit on construction to ensure that state of show_focus and
show_parameters are set correctly when toggling units. Also refactor
code for showing/hiding focus controls in skin to a callback connected
to show_parameters so it gets triggered by clicking the show_parameters
button in skins as well as effectFocusButton on the controller.