hold loop_in/out to move loop boundaries; catch loop with reloop CO
When a loop is enabled, let users move the loop in/out points by holding
down the loop_in/out buttons. The new point will be set when the button
is released. Previously, the loop had to be disabled before setting new
loop boundaries outside of the current loop.

Change behavior of "reloop" ControlPushbutton:
* no loop active, playhead before the loop in point: enable loop but
do not jump to it
* no loop active, playhead inside loop: enable loop
* no loop active, playhead after loop: seek to loop in point, enable
loop, and stop
* loop enabled: seek to loop in point, enable loop, and stop
This makes the button useful for setting up loops whether they are being
used to mix out a track or mix in a track.