remove "reloop" CO, make "reloop_toggle" catch upcoming loops
Behavior of "reloop" CO was good for pushing loop encoders with
shift, but was not good for skins or controller buttons because
there was no way to exit a loop with it. The use case of preparing
to mix in a loop is still covered by reloop_andstop, which will
be used as the right click control for skin reloop buttons and
is suggested as the shift function of reloop buttons on
controllers.