make LoopingControlTest.LoopMoveTest pass
There was an issue with the loop in and out points being rounded
down to even samples with the extremely high BPM used by the test
before. That issue should be fixed by refactoring the engine to use
floating point sample positions, but that is beyond the scope of
the current looping changes (GitHub PR #1187). It seems this issue
only comes up when the BPM is so high that rounding beats to the
nearest frame could produce a different result.