Deere: re-remove effect chain enable switch
It adds an unnecessary layer of complexity to a complex system. As far
as I know, only one controller, the Native Instruments Kontrol S4, has a
button specifically for it. Notably, no other NI controllers have that
button. The usability of the skin should not be reduced for everyone for
a single controller, and I don't think it is worth adding an option to
the skin settings menu either.