measure beatloop_size when setting manual loops with quantize
Plus:
* make loop_double/halve always update beatloop_size
* remove beatloop_double/halve ControlPushButtons

Now, beatloop_size can only be out of sync with the loop in two cases:
* a new track is loaded with a loop of a different size
* a manual loop is set using loop_in/out without quantize

It is no longer possible to resize manual loops that are not quantized
with loop_double/halve. As discussed on PR #1187, that is not very
useful behavior. It is still possible to resize manual loops that are
not quantized by setting loop_scale. That will not update beatloop_size.