Revert "apply effects to PFL channels before mixing into headphone mix"
This reverts commit ee10191015ac39066a21ffebd79cbe56013cf617.
Applying effects for both the PFL and master mix caused effects to
be processed twice per channel per engine callback. Effects keep an
internal state for each channel, so manipulating that twice per
callback was causing some very nasty issues when a deck was assigned
to an effect unit and the headphone mix. For example, the reverb
effect fed back into itself infinitely and produced a screetching
sound within a few seconds.

For Mixxx 2.1, to preview effects in headphones, the effect unit
will have to be assigned to the headphone channel and the deck
will have to be assigned to the headphone mix. Having to press two
separate buttons is not a great user experience, especially
considering controllers do not have buttons to assign effect units
to headphones. However, automatically applying effects to the
headphone mix in addition to the master mix would require
complicated changes to the engine with a high risk for regression,
so that is being postponed for the future.