Add crate filter in library search
Add CrateFilterNode in library/searchquery.h witch is responsible
for handling the crate filtering. Modify searchqueryParser.cpp
to handle the crate keyword and create a CrateFIlterNode.
Finally add a few functions in cratestorage.cpp to handle
the SQL queries needed for the filter to work.
Also change searchqueryparsertest.cpp due to the
new SearchQueryParser constructor.