Components: refactor Button input handling
remove the old onlyOnPress Button property. When set to false, this
would toggle the value of a Control on button up and button down, which
is not really the correct behavior. Instead, let Buttons have a type,
with the default type being push. The button types are:

push: set inKey to 1 on button press and 0 on button release
toggle: toggle value of inKey on button press
trigger: set inKey to 1 on button press
powerWindow: like toggle, but toggles the value of inKey again on button
up when long pressed

The ComponentContainer.shift() and ComponentContainer.unshift()
functions automatically take care of setting the inKey of push type
Buttons to 0 on shift/unshift to ensure that it does not matter whether
shift or the Button is released first.