add separate booth output with mics optionally mixed in
This replaces the option to exclude the microphone inputs from the
master output. That was sufficient for broadcasting users, but there was
no way to have a master output with the mics and a booth monitor signal
without the mics for live use. Previously that required plugging a mic
into an external hardware mixer, but now it is possible to do with Mixxx
and any sound card with enough output channels.