avoid modifying inputs to ChannelMixer::applyEffectsAndMixChannels(Ramping)
This was causing popping sounds on the master output when routing
microphones to headphones because
ChannelMixer::applyEffectsAndMixChannelsRamping was applying gain to the
channel buffer. To work around this, move the application of gain to
EngineEffectsManager::processPostFaderAndMixRamping, where the channel
input buffer is copied into one of EngineEffectsManager's temporary
processing buffers. Using EngineEffectManager's buffer avoids allocating
a temporary buffer for every EngineChannel.

Also, remove the unused ChannelMixer::mixChannels(Ramping) functions. If
their behavior is desired in the future, nullptr can be passed as the
EngineEffectsManager* argument to
ChannelMixer::applyEffectsInPlaceAndMixChannels(Ramping).