track enable state of effects & chains for each input and output
This is required to send the intermediate enabling/disabling states for
every combination of input and output. Previously, when an effect was
processed on multiple channels, only one of them would get the disabling
signal when turned off. This caused the Echo effect to play the sound
last stored in its buffer when it was disabled with multiple inputs
going to it. Echo's buffer would be cleared for only one of the input
channels.