Use QSharedPointers instead of std::unique_ptr and add missing connect call
BroadcastProfile notifies name changes to BroadcastSettings with Qt's signals &
slots system. However Qt is not compatible with the std::unique_ptr smart
pointer. A refactor of BroadcastProfilePtr is required to make it compatible.
std::unique_ptr has been replaced with QSharedPointer, as this type is easy
to pass around in the code while keeping a reference to the original pointer
and deleting it when not used anymore.