do not let samplers be routed to effects
This required effects to maintain state for every sampler going
to both master and headphone outputs. With 64 samplers, memory use
got ridiculous for effects that allocate a buffer for each channel
like Echo. In Mixxx 2.0, no skins showed the effects routing
switches for samplers, so I do not think users will miss this.
Finding space for these switches in the Mixxx 2.1 skins has been
a challenge, so this will be relieving for skin design.