ctlra: add emtpy enumerator class
This commit adds the enumerator class for Ctlra devices.
The main ctlra context is owned by the Enumerator class,
as this is where the device hotplug callbacks reside.
When a new device is hotplugged, it can be added/removed
from the list of controllers in the UI panel easily, as
that is the class that owns the context.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>