ctlra: refactor CtlraReader, add event print func
This commit refactors the CtlraReader to operate correctly,
and adds an event printing function to show what events are
handled by Mixxx's Ctlra instance.

All devices are accepted, and all events are routed to the
same handling function, which obviously doesn't scale to
having multiple controllers use different mappings. That will
be solved when the CtlraController class is introduced.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>