ctlra: move event_func to CtlraController class
This commit moves event handling into the CtlraController class,
enabling easy lookup of the Controller properties, like which event
handling script to use, or what to do with the events.

Deriving a class from CtlraController would enable various specializations
to exist. For example, if a JS Mixxx integration is required, it could
derive a JSCtlraController, or if TCC was desired, it would derive the
TCCCtlraController.

If a specific devices *requires* C/C++ code for performance reasons, it
could derive a C/C++ class to handle events or provide a very-high
performance backend for a specific device. Thinking here is that HD
screens should be rendered using C/C++ code, and hence this code will
require access to the device itself to blit the pixels over the cable.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>