ctlra: build depends/features link ctlra better
This commit adds the Ctlra library to the Mixxx depends.py file,
and as such allows the build system to be aware of the library to link.

Given that Ctlra doesn't have a .pc file (yet), the library path to link
with is retrieved from an environment variable.

In order to build Mixxx with Ctlra support, do the following steps:
- git checkout https://github.com/openAVproductions/openAV-Ctlra ctlra
- cd ctlra
- mkdir build && cd build
- cmake ..
- make -j4
- export CTLRA_PATH=`pwd`

Then using the *SAME* terminal, cd <mixxx>, checkout the ctlra branch,
and type "scons ctlra=1". Other flags (-j, optimize etc) can be used as
normal.

Any questions on compiling, please contact me.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>