Rename now avaliable. Fix typos
Along with rename the functions used to determine if a crate can be
named like the user wants to are now more robust. (relying on
crateId instaed of crateName)

I can't seem to be able to crash the application by having a crate
with a name that it shouldn't have. (as it would happen a lot in my
tests)