Deere: remove minimal deck controls mode
This was a quick hack around the clutter of the old looping UI. Now that
all controls fit into a reasonable amount of space and the bottom
splitter has been removed, get rid of this option. On many occasions I
found myself confused why I couldn't access a control in minimal
controls mode then annoyed that I had to open the skin settings menu to
get to what I wanted.