Fix miniscrollbar hint letters
The problem was that the `label` for the tree items
did not match the `data` of the items, EXCEPT for the
tracks. In the case of crates and playlists, only the
`label` was the human readable version, whereas the
`data` was the playlist/crate ID in the sqlite database.

It could be that to enable drag/drop from the miniviews
that we will need to modify the tree items in for the
mini track view so that their data is also their track ID.