remove persistence from EffectUnit routing switches
I thought this would be useful before when working on #1103, but
experience has shown it is annoying. I want effect unit 1
assigned to deck 1, unit 2 to deck 2, unit 3 to deck 3,
and unit 4 to deck 4 on startup regardless of how the routing
switches happened to be when I last shut down Mixxx.