On untagged queries us token as crates name.
Currently it is only possible to use the advanced search "crates:" to search
for tracks in specific crates. Untagged search queries will only be matched to
the track, but not to the crates the track is in.

This change allows you to use crates like tags and query them unspecific.