Echo effect: reduce size of buffers
The Echo effect referred to the
mixxx::AudioSignal::SampleRate::max() constant, equal to 192 kHz,
to determine the size of the buffers it allocates. This resulted
in a huge waste of memory.

There is no reason Mixxx should support 192 kHz, even if a user's
sound card does. It provides no benefit and is likely to introduce
introduce intermodulation distortion when playing ultrasonic
frequencies. Refer to
http://xiph.org/~xiphmont/demo/neil-young.html
for details.