Echo: reduce maximum delay time back down to 2 seconds
There is a tradeoff between the maximum delay time and the memory required
by the effect. The higher the maximum delay time, the lower the minimum
BPM that can use the full range of 2 beats. The cutoff point is arbitrary.
The default minimum BPM for analysis is set at 70, so I think it is fairly
unlikely that a user would use the Echo effect with the maximum delay time
of 2 beats with a track less than 60 BPM.