load EffectStates only for enabled input channels
Previously, effect states were preallocated in the main thread for every
combination of input and output channel, then a pointer to the effect
was passed to the audio engine thread by the effect MessagePipe FIFO.
For effects that store longer buffers like Echo, this added up to a lot
of memory. This commit adds a new EffectState abstract base class that
each effect needs to subclass. These EffectStates are allocated and
destroyed in the main thread whenever an input channel is enabled or
disabled for an effect unit. When a new effect is loaded, EffectStates
are allocated for all channels enabled at that time.

The new EffectState class uses the new mixxx::AudioParameters class to
take a step towards removing the assumption of stereo buffers from the
engine. This also paves the way for allocating buffers of the
appropriate sizes throughout the engine instead of relying on the huge
MAX_BUFFER_LEN constant.

Lastly, this commit replaces the old fasioned
EffectProcessor::EnableState enum with the EffectEnableState enum class
in a new header file.

TODO: fix effects tests