decouple EffectStatesMapArray from EffectChain
This fixes a crash deleting EffectStates after starting up with an
effect unit assigned to multiple input channels.
EffectChain::enableForInputChannel kept reusing the same array for
EffectStates and putting a pointer to this on the effects MessagePipe.
That assumed that the engine had a chance to process the message to load
the states by calling EffectProcessorImpl::loadStatesForInputChannel
before the next call to EffectChain::enableForInputChannel. On startup,
before the engine starts, this assumption is not true. The result was
that EffectProcessorImpl ended up storing pointers to the same
EffectStates for multiple input channels. This caused a crash when
~EffectProcessorImpl tried to delete the EffectStates for the second
input channel because the EffectState had already been deleted when
deleting the EffectStates for the first input channel.

Fixes https://bugs.launchpad.net/mixxx/+bug/1740531