tweak scratching behavior
Behave differently when the platter is spinning forward vs backwards
but not being touched. Effectively this changes causes the track to
stop scratch mode and start playing sooner when moving foward vs
backwards. This allows better spinbacks, and more natural cueuing when
seeking backwards with the platter and for cleaner exits when
scratching when pushing the platter forward at the end of a scratch.