American Audio VMS2 & VMS4/4.1: unmap volume faders
These devices combine a MIDI controller and hardware mixer.
Unfortunately there is no way to decouple these different functions
except for bypassing the VMS' equalizers. In Mixxx 2.1, the
umixed Deck 1-4 outputs were changed so they are now affected by the
deck volume faders. This was done because effects are now processed
after the volume faders. Sending the output signals to the Deck 1-4
outputs before the volume fader would make effects unusable with
external mixers. (Alternatively, it would be possible to do duplicate
effects processing before the volume faders for the Deck 1-4 outputs
like what is done for the Headphones output, but that would
complicate the mixing engine code and of course the software
volume fader could not be used to cut input to effects.)