undo volume fader attenuation on input to Bitcrusher and redo it on output
The character of this effect changes with the input level and effects
are processed after the volume fader. So, to preserve the character of
the effect as the volume fader is lowered, undo the attenuation from the
volume fader on the input for this effect, then reapply it on the
output.