A&H Xone K2/K1: rewrite 4 deck mapping with Components
This is required for reusing the code for a 2 deck + 2 effect unit
mapping. The Components library replaces the use of the Latching Layers
setting in the K2's firmware, which is not available with the K1. So
this also had the benefit of making the mapping compatible with the Xone
K1.