Components: add capability for multicolor buttons with EffectUnit
The EffectUnit constructor can be passed a 4th argument now to specify
different colors for different states. The argument should be an object
with three properties: focused, unfocused, and focusChooseMode. The
EffectUnit will set a "color" attribute of the Buttons to the values
passed for each state. It is up to the controller mapping to reimplement
components.Button.prototype.send to take the color attribute of the
Button into account as appropriate for that controller.