sort controller mappings in user folder first then system folder
This was how it was done in 2.0. It is not self-explanatory and
kinda clunky, but it's better than the current situation where
they are all sorted alphabetically and users get confused why
they see two identically named mappings next to each other.

This behavior was broken by commit effbb06 because it removed the
QMap of mappings indexed by a QString for the file path.