Kontrol S4: implement effect focusing
When no effect is focused, the effect knobs control the metaknobs and
buttons control the effect enable buttons. When an effect is focused,
the knobs control the parameters of the focused effect and the buttons
control its button parameters.

Holding the button below the mix knob then pressing one of the effects
buttons focuses an effect. If the button for the currently focused
effect is pressed, the effect is unfocused and the knobs go back to
controlling the metaknobs. However, I don't have access to the hardware
right now, so I can't test if that button input is mapped correctly.