WEffectSelector horribly ugly hack to hide checkmarks on macOS only
macOS displays a wide checkmark on the left of the drop down menu
that displaces text and makes it get cut off on the right side.
The checkmark icon can be hidden by styling WEffectSelector::indicator,
but the space it occupies cannot be easily changed. So, set a negative
left margin for WEffectSelector QAbstractItemView. However, that
causes the text to get cut off on the left on KDE, so here is a horribly
ugly hack to set the QSS margin from C++ in a platform dependent way.