fix crash pressing Apply in DlgPrefLV2
when Mixxx is loaded with an effects.xml file that has any empty
effects chains

fixing https://bugs.launchpad.net/mixxx/+bug/1773097

This is just a quick hack on top of an overcomplicated
architecture. EffectChain and EffectChainSlot need to be merged
before adding any more features to the effects system.