remove Reverb Dry/Wet parameter and restore old mixing behavior
2b72491ee2a0bd761e0d68600b62ebbb3b89dc6c changed the Reverb effect so
that the dry signal was not passed through when the effect unit was in
D+W mode. A new dry/wet parameter was added to compensate for this in
fd25553f2905c07165980084914ed648fe6dfa81 so there would still be a way
to adjust the relative prominence of Reverb in a D+W chain. However, no
change to the mixing of Reverb was needed in the first place. Simply
keeping the dry signal with a Send parameter works in both D/W and D+W
modes.