Echo effect: do not artificially clamp internal buffer
It is up to the user to control the level with the chain mix knob
and Echo's send knob. It can be surprising to accidentally distort
the signal when an Echo effect is in a chain. Digital signals are
not clamped elsewhere in Mixxx before they are sent to the audio
interface, so I do not think they should be here. If the user does
want distortion, we can provide that in the future with a
distortion effect (refer to
https://bugs.launchpad.net/mixxx/+bug/1706444 ).