send report IDs of HID output packets
Before, HID controller mappings were improperly relying on an
implementation detail of hidapi by putting the report ID as the
first byte of the packet and not specifying a report ID. hidapi
requires a report ID specified as the first byte sent to
hid_write. For devices that only have a single output report type,
hid_write expects a 0 as the first byte. So, Mixxx prepends a
0 when the script does not specify a report ID.

Windows requires the leading 0 byte for devices that only have a
single output report type, so hidapi sends the data from Mixxx
to Windows. However, on Linux, libusb does not require prepending
a report ID of 0 when the device only uses one output report type,
so hidapi strips the leading 0 byte before sending the data to
libusb. Thus, when Mixxx prepended a 0 byte because the scripts
did not specify a report ID, it happened to work on Linux but did
not work on Windows.