Use shared_ptr for AudioDestination/AudioSource interfaces.
This adds safety against use-after-free to the interaction between
EngineMaster, SoundManager, EngineSideChain, VinylControlProcessor,
EngineDeck, EngineMicrophone, and EngineAux.

Note that this effectively reverts
db3b503ea954cb5fbf54a731a3c0f30a25fc10dc. The reason this was done
initially was to avoid coupling between SoundManager and
EngineMaster. Ideally neither of them should know about each other,
and they could communicate via interfaces alone. They are currently
coupled (SoundManager has a reference to EngineMaster) because of the
callback, but I've always wanted to remove that so they are totally
independent.

PlayerManager (which could have been named EngineManager) as I
originally imagined it, is where EngineMaster and SoundManager are
bound together, since PlayerManager has a reference to both; for
example, when a deck is added, it's added to EngineMaster and
registered with SoundManager in PlayerManager. That's why the logic to
register EngineMaster's non-EngineChannel AudioInputs/AudioOutputs
originally went in PlayerManager. Thinking of PlayerManager as
EngineManager, this kind of makes sense.

In this PR, the logic can no longer live in EngineMaster for the
simple reason that EngineMaster does not have a shared_ptr to itself,
and so it can't call SoundManager::registerXXX(..., this) anymore. We
could plumb a shared_ptr into EngineMaster, but that creates a
reference leak.