Build Mixxx on Ubuntu 18.04 via AppVeyor.
Travis only supports Ubuntu 14.04 (Qt 5.2), so we cannot use it for newer Qt features.

Also disables the x86 build for Windows, since we only get one
concurrent build on AppVeyor and it's rare for an issue to affect x86
and not x64. (It's also rare to see a user on a 32-bit operating system
these days.)