Move GuiTick processing from VSyncThread to WaveformWidgetFactory.
GuiTick is delivering more and more render signals to widgets in the main
thread, and when emitted from the VSyncThread, we're putting a heavy load on the
Qt event queue. When these signals are emitted from
WaveformWidgetRenderer::render(), they are emitted from the main thread, and so
can be delivered directly rather than queued.