copy play state closer to the position sync operation
Previously the track would start playing as soon as it was loaded but
before the position was synced. Some audio would leak out on occasion.
Moving these operations closer together reduces this possibility. To
completely eliminate it, the seek and the play will need to happen in
the engine thread together. Not sure if this will be necessary.