track the filename we used for broadcast profiles
Previously the filename was always derrived from the profile name which
causes problems if the names differ. For example, it was impossible to
remove a profile that had a `.` in the original profile name, after a
mixxx restart, unless you renamed the profile to the original name then
attempted the removal.