remove call to qt_x11_wait_for_window_manager
This was previously guarded by an #ifdef Q_WS_X11, but switching
to Q_OS_LINUX caused a link error:

/usr/bin/ld: lin64_build/mixxx.o: in function `MixxxMainWindow::MixxxMainWindow(QApplication*, CmdlineArgs const&)':
/home/uk/Projects/Mixxx/mixxx/src/mixxx.cpp:187: undefined reference to `qt_x11_wait_for_window_manager(QWidget*)'

It turns out that Q_WS_X11 is no longer defined with Qt5, yet it
does not seem that building Mixxx without the call to
qt_x11_wait_for_window_manager has caused any issues, so I presume
it is safe to remove it now. This was the only place in Mixxx
where Q_WS_X11 was referenced.