[WIX] Again, pull latest translations from Transifex after manually filling strings from translations memory. They were lost in a previous commit 343246a21e03f09990d5522bae5001174a622c26
``tx pull -r mixxxdj-windows-installer.mixxxpot -a --minimum-perc=1``

The template and the translations became out-of-sync, since a manual push to TF following https://github.com/mixxxdj/mixxx/pull/1899
I do not exactly know what the issue was, probably it does not work well that strings in the language files were manually changed, but not the timestamp.

Anyway, now the template and the translations are n-sync again. Back to normal. In the future , we should NOT accept manual edits to translation files.
Either use the TF online editor or one of the many available text translation softwares, see https://mixxx.org/wiki/doku.php/internationalization#offline_translation_software