Restructure the flow of rendering QOpenGLWidgets.
Summary:
* Remove responsibility for swapping QOpenGLWidgets from the VSyncThread.
* Introduce BaseQOpenGLWidget, a base class for QOpenGLWidget waveforms.
* Inherit all QOpenGLWidget waveforms from BaseQOpenGLWidget.
* Use aboutToCompose/frameSwapped signals from QOpenGLWidget to determine when
  to render widgets.
* Add the same logic from BaseQOpenGLWidget to WSpinny.

After this change the VSyncThread's purpose is to send render ticks to the main
thread at the user's configured FPS. It still communicates with the main thread
via semaphore, so it can't flood the main thread with signals. A render signal
from the VSyncThread causes all QOpenGLWidgets to set a boolean "render on next
tick", which when QOpenGLWidget::aboutToCompose fires next, kicks off a blocking
render of the widget.

One important function of the VSyncThread is that it estimates when the next
vsync is. We use this to smoothly interpolate each deck's playposition so that
the position we render on the screen at the next swap is where the deck *was*
or *will be* at the moment the audio comes out the speakers, assuming the deck
continues forward linearly at its last measured rate. This prevents the waveform
from "jerking" caused by the interplay between the waveform render period and
the audio callback period.

To get this equivalent behavior in the new setting, I added an estimator for the
time until the next swap by measuring the time between a QOpenGLWidget receiving
an aboutToCompose signal from Qt's backing store until the time it receives a
frameSwapped signal indicating the widget is now visible. Each widget keeps an
exponentially weighted moving average of this estimate, and uses it
to estimate the time until swap, which is fed into the playposition
interpolation logic in VisualPlayPosition.