Work around performance issue with Qt 5.12 and PreviewButtonDelegate. Launchpad Bug #1812763
After upgrading to Qt 5.12, Mixxx idle CPU usage is high (99%) on
macOS. Profiling shows LibraryPreviewDelegate::paintItem as one of the hottest
functions. It appears to be related to calling QPushButton::setGeometry for each
row of the library. My workaround relies on the fact that we only need the
QPushButton size to be correct, not its position.

On macOS 10.13.6, CPU usage drops from 99% to 50% with this change
(and with #1994, idle usage drops to 15%).