The Crossfader is not always used, also on most controllers (such as the Hercules P32 DJ) its knob does not lock in the middle position to avoid unwanted changes.
As a result it is possible that the Crossfader knob is moved involuntarily
away from its middle position during normal controller operation, causing
unwanted (deck) volume imbalances.

This patch adds an option (checkbox) to the Mixxx Graphical User Interface
so that it is possible to disable the Crossfader functionality (controller
mapping).

When enabling the AutoDJ functionality make sure to also re-enable the
Crossfader if it has been previously disabled (thanks to Daniel Schürmann
for raising this issue).

The translation for the label of the new GUI checkbox is provided for the
Italian language only (and English defaults).

The logic for setting the various other Crossfader options is also fixed
and made consistent to avoid bugs.

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 res/translations/mixxx_en.ts                   |   19 +++---
 res/translations/mixxx_en_GB.ts                |   19 +++---
 res/translations/mixxx_it.ts                   |   21 ++++--
 src/engine/enginemaster.cpp                    |   10 +++
 src/engine/enginemaster.h                      |    1
 src/library/autodj/autodjprocessor.cpp         |   14 ++++
 src/library/autodj/autodjprocessor.h           |    5 +
 src/preferences/dialog/dlgprefcrossfader.cpp   |   77 ++++++++++++++++++-------
 src/preferences/dialog/dlgprefcrossfader.h     |    4 +
 src/preferences/dialog/dlgprefcrossfaderdlg.ui |    9 ++
 10 files changed, 136 insertions(+), 43 deletions(-)