make better use of intro/outro markers with AutoDJ
and add a checkbox to revert to the old behavior that does not use the
new intro/outro markers at all. With the box checked (which is the
default), AutoDJ uses the outro of the outgoing track or the intro of
the incoming track for the transition time, whichever is longer. If only
one track has these marked, use that track's marked transition time. If
neither track has the intro/outro range marked, fall back to the fixed
number of seconds specified by the spinbox. If users do not want to
interrupt the workflow they are accustomed to when upgrading to Mixxx
2.3, they can simply uncheck the new box.