Fix key rounding code, returning max + 1.
Originally the chords and the chroma data have 3 bins per key and center at 1.
This means the first correlation matches at 0. To compensate this the chroma
data is shifted on step more. This way we have C wihout wrap around at 0, 1
and 2 and can use integer math to get the final key value.
This was tested with a B minor chord and a 440 Hz (A) and 523,25 Hz (C) sinus.