Roland DJ-505: Add track preview using the Browser Knob
This allows to quickly listen to the songs in PreviewDeck1.

When the Browser Knob is pressed (i.e. used as a button),
the currently selected song is loaded into the preview deck and playback
is started. Pressing it again will stop the playback.

Turning the Browser Knob while holding it down allows to quickly seek
through the track loaded in the preview deck.

The shifted behavior (i.e. selecting a sidebar item on knob turns,
expanding/collapsing it on knob presses) will not be changed by this
commit.

Previously, pressing the knob would always attempt to collapse/expand a
sidebar item. If a track was selected instead, this resulted in a no-op.