Roland DJ-505: Fix Pitch Play mode
Previously, the Pitch Play mode was setting the "key" control for the
channel to the sum of "file_key" and the modifier value. This was
error prone since key could become negative (depending on the track),
which would then result in no key change.

Instead, the "pitch_adjust"control is now used, which is not only a
relative, signed value and works much better, but also avoids reading
the "file_bpm" control and the additional calculation with that value.