Move version into own file and use version file in default.nix
Due the nature of nix, using "git describe" will always waste a
copy of the source folder and is quite slow.
Since nix string functions are very limited and extracting the
version is difficult, the best compromise was to move the version
into a small file as a single source of truth.