parented_ptr: Assert parent at destruction instead of construction
Previously, `parented_ptr` asserted on construction that the pointed-to
object has a parent. However, sometimes one wants to create an object
and assign it to a parent later. The
[recommendation](https://www.mixxx.org/wiki/doku.php/coding_guidelines#pointer_object_lifetime_ownership)
was to create a `unique_ptr` `p` instead, later assign the object to a
parent and then create a new variable using `make_parented(p.release())`.

This has however several drawbacks:

* There is a short period where the object is already assigned to a
  parent but still owned by a `unique_ptr`. This can lead to a double
  free.
* There are two pointers that and it depends on the code location which
  one is the correct one to access the object.
* It is inconvenient to use, because one has to write an extra line
  where the `unique_ptr` is released.

With this commit, `parented_ptr` asserts that the pointed-to object has
a parent at the **destruction** of the `parented_ptr`. That means, that
one can store the object using a `parented_ptr` a priori and there will
be a warning (from the failing `DEBUG_ASSERT`) in case any object is
leaked.

Optionally, in a future PR, this behaviour can even be changed so as to
delete the object that is going to be leaked. That is, the
`parented_ptr` destructor wouldn't only `DEBUG_ASSERT` that the
pointed-to object has a parent, but it would also delete it otherwise.

Additional changes:

* Made `parented_ptr` implicitly convertible to a raw pointer. The
  reasons why the standard library types are not convertible to a
  raw pointer don't apply to `parented_ptr`. Regarding the insufficient
  usage of `parented_ptr` in Mixxx at the moment, making it implicitly
  convertible will make its adoption easier because less code has to
  be changed. (No `.get()` everywhere.)
* Added a constructor taking `nullptr_t`.
* Added `noexcept` where applicable.
* Formatted the code using `clang-format`.