Roland DJ-505: Remove LOOP mode
In its current form, the LOOP mode is not really useful. It maps the
pads to normal beatloops of different sizes - similar to the ROLL mode.
I'm not using it at all, since the Loop Sections is much more
convenient. For people coming from Serato, this is quite unintuitive
since it behaves completely different from the expected Saved Loop mode.

Unfortunately, Mixxx does not have the ability to save loops yet [1].
However, if this feature is added at some point, we can easily add a
proper loop mode. By removing it now (before it becomes part of a Mixxx
release) we can avoid having to document the previous behaviour forever.

[1] https://bugs.launchpad.net/mixxx/+bug/692926