don't identify PortAudio devices by PaDeviceIndex in configuration
PortAudio provides no guarantees that the PaDeviceIndex for any
device will be stable across restarts of Mixxx. Abusing the
PaDeviceIndex as a persistent identifier caused Mixxx to falsely
claim that audio interfaces were not available and annoyingly
require the user to reconfigure all their sound I/O even when
nothing about the available sound hardware actually changed. It
is the responsibility of the sound API to provide persistent
names for devices to PortAudio.