Fix shift behaviour for ComponentContainers
For advanced controller mappings, it can be desireable to check the
current shift state of a component container by reading its `isShifted`
attribute. This attribute is defined as follows:

    isShifted: false,
    shift: function () {
        this.forEachComponent(function (component) {
            if (typeof component.shift === 'function') {
                //[...]
                component.shift();
            }
            // Set isShifted for child ComponentContainers forEachComponent is iterating through recursively
            this.isShifted = true;
        });
    },
    unshift: function () {
        this.forEachComponent(function (component) {
            if (typeof component.unshift === 'function') {
                // [...]
                component.unshift();
            }
            // Set isShifted for child ComponentContainers forEachComponent is iterating through recursively
            this.isShifted = false;
        });
    },

Unfortunately, the comment in the source code is wrong, because `isShifted` will
*not* be set recursively. In fact, it will only ever be set if the
`shift` method of a component container is called directly. When
interating thought components recursively, `forEachComponent` will only
call `shift` for the components inside a component container, but not
for the container itself:

    var applyOperationTo = function (obj) {
        if (obj instanceof Component) {
            operation.call(that, obj);
        } else if (recursive && obj instanceof ComponentContainer) {
            obj.forEachComponent(operation);
        } else if (Array.isArray(obj)) {
            obj.forEach(function (element) {
                applyOperationTo(element);
            });
        }
    };

Hence, `isShifted` will not be set recursively.
To fix that, this PR introduces an additional function
`forEachComponentContainer` that will be used by the
`shift()`/`unshift()` methods to propagate the shift state to child
containers recursively.